requirements.txt
# requirements.txt
pandas
uvicorn
fastapi
jinja2
mysql-connector-python
boto3
# Add other dependencies as needed
